Rename modules to containers
