+++ src/components/Molecules/List/Events/Events.tsx
@@ -5,7 +5,7 @@ import { translate, TranslationFunction } from 'react-i18next';
 import EventsHeader from '../../../Organisms/EventsHeader/EventsHeader';
 import EventCard from '../../../../EventCard/EventCard';
 import compose from 'recompose/compose';
-import { IMeetupEvent } from '../../../../modules/EventsContainer/EventsContainer.interfaces';
+import { IMeetupEvent } from '../../../../containers/EventsContainer/EventsContainer.interfaces';
 import { NextEventsBlock } from './NextEventsBlock';
 import TrainingsSection from '../../TrainingsSection/TrainingsSection';
