+++ src/containers/EventsContainer/EventsContainer.helpers.ts
